refactor(frontend): move briefing card measurement math into helpers
scrollByStep always requested smooth scrolling, so the scroll arrows kept
animating for readers who had asked for reduced motion.

The measurement math and scroll-edge maths move to helpers.ts, leaving the
hook to own state and the observer lifecycle. COLLAPSED_ROWS moves with
them, since both the helper and the card need it.
